+++ src/adapter/http/tests/tests/tests_push/scenarios/mod.rs
@@ -13,6 +13,7 @@ mod scenario_existing_dataset_fails_as_server_advanced;
 mod scenario_existing_evolved_dataset;
 mod scenario_existing_up_to_date_dataset;
 mod scenario_new_dataset;
+mod scenario_new_empty_dataset;
 
 pub(crate) use scenario_aborted_write_of_new_rewrite_succeeds::*;
 pub(crate) use scenario_aborted_write_of_updated_rewrite_succeeds::*;
@@ -20,3 +21,4 @@ pub(crate) use scenario_existing_dataset_fails_as_server_advanced::*;
 pub(crate) use scenario_existing_evolved_dataset::*;
 pub(crate) use scenario_existing_up_to_date_dataset::*;
 pub(crate) use scenario_new_dataset::*;
+pub(crate) use scenario_new_empty_dataset::*;